+++ src/server/graphql/isSubscriptionPayload.js
@@ -1,0 +1,3 @@
+const isSubscriptionPayload = (payload) => payload.query.startsWith('subscription')
+
+export default isSubscriptionPayload