+++ src/universal/components/ExternalLink.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ExternalLink = ({className, copy, href}) => (
+  <a href={href} rel='noopener noreferrer' target='_blank' title={copy} className={className}>
+    {copy}
+  </a>
+)
+
+export default ExternalLink