+++ src/universal/components/MeetingHelp/HelpMenuLink.js
@@ -1,0 +1,8 @@
+import styled from 'react-emotion'
+import ExternalLink from 'universal/components/ExternalLink'
+
+const HelpMenuLink = styled(ExternalLink)({
+  textDecoration: 'underline'
+})
+
+export default HelpMenuLink