+++ src/universal/components/RetroDiscussPhase.js
@@ -9,16 +9,19 @@ import IconLabel from 'universal/components/IconLabel'
 import ui from 'universal/styles/ui'
 import appTheme from 'universal/styles/theme/appTheme'
 import StyledFontAwesome from 'universal/components/StyledFontAwesome'
+import LabelHeading from 'universal/components/LabelHeading/LabelHeading'
 import ReflectionCard from 'universal/components/ReflectionCard/ReflectionCard'
 import MeetingAgendaCards from 'universal/modules/meeting/components/MeetingAgendaCards/MeetingAgendaCards'
 import findStageAfterId from 'universal/utils/meetings/findStageAfterId'
 import EndNewMeetingMutation from 'universal/mutations/EndNewMeetingMutation'
-import {withRouter} from 'react-router-dom'
 import type {RouterHistory} from 'react-router-dom'
-import ScrollableBlock from 'universal/components/ScrollableBlock'
+import Overflow from 'universal/components/Overflow'
+import {meetingGridGap, meetingGridMinWidth, meetingVoteIcon} from 'universal/styles/meeting'
+import plural from 'universal/utils/plural'
+import {withRouter} from 'react-router-dom'
 import {connect} from 'react-redux'
 import type {Dispatch} from 'redux'
-import {meetingTopicPhaseMaxWidth, meetingVoteIcon} from 'universal/styles/meeting'
+import DiscussHelpMenu from 'universal/components/MeetingHelp/DiscussHelpMenu'
 
 type Props = {|
   atmosphere: Object,
@@ -29,10 +32,33 @@ type Props = {|
   team: Object
 |}
 
+const maxWidth = '114rem'
+
+const HeaderContainer = styled('div')({
+  margin: '0 auto',
+  maxWidth,
+  padding: '0 1.25rem'
+})
+
+const LabelContainer = styled('div')({
+  margin: '0 1.25rem',
+  padding: '0 0 .625rem'
+})
+
 const DiscussHeader = styled('div')({
   alignItems: 'center',
   display: 'flex',
   margin: '0 0 1.25rem'
+})
+
+const ColumnsContainer = styled('div')({
+  display: 'flex',
+  flex: 1,
+  margin: '0 auto',
+  maxWidth,
+  overflowX: 'auto',
+  padding: 0,
+  width: '100%'
 })
 
 const TopicHeading = styled('div')({
@@ -70,39 +96,30 @@ const PhaseWrapper = styled('div')({
   overflow: 'hidden'
 })
 
-const ReflectionSection = styled('div')({
-  borderBottom: `.0625rem solid ${ui.dashBorderColor}`,
+const ReflectionGrid = styled('div')({
+  display: 'grid',
+  gridGap: meetingGridGap,
+  gridTemplateColumns: `repeat(auto-fill, minmax(${meetingGridMinWidth}, 1fr))`
+})
+
+const Column = styled('div')({
   display: 'flex',
   flex: 1,
-  flexShrink: 0,
   flexDirection: 'column',
-  margin: '0 auto',
-  maxWidth: meetingTopicPhaseMaxWidth,
-  minHeight: 200,
   width: '100%'
 })
 
-const ReflectionSectionInner = styled('div')({
-  padding: '0 2.5rem 1.25rem'
+const TaskColumn = styled(Column)({
+  borderLeft: '.0625rem solid rgba(0, 0, 0, .05)'
 })
 
-const ReflectionGrid = styled('div')({
-  display: 'grid',
-  gridGap: '1.25rem',
-  gridTemplateColumns: 'repeat(auto-fill, minmax(240px, 1fr))'
-})
-
-const TaskSection = styled('div')({
-  display: 'flex',
-  flex: 1,
-  height: '10.25rem',
+const ColumnInner = styled('div')({
+  padding: '.625rem 1.25rem 1.25rem',
   width: '100%'
 })
 
 const TaskCardBlock = styled('div')({
   margin: '0 auto',
-  maxWidth: meetingTopicPhaseMaxWidth,
-  padding: '1.25rem 2.5rem',
   width: '100%'
 })
 
@@ -140,45 +157,59 @@ const RetroDiscussPhase = (props: Props) => {
   return (
     <React.Fragment>
       <PhaseWrapper>
-        <ReflectionSection>
-          <ScrollableBlock>
-            <ReflectionSectionInner>
-              <DiscussHeader>
-                <TopicHeading>
-                  <span>{'“'}</span>
-                  {`${title}”`}
-                </TopicHeading>
-                <VoteMeta>
-                  <VoteIcon name={meetingVoteIcon} />
-                  {voteCount}
-                </VoteMeta>
-              </DiscussHeader>
-              <ReflectionGrid>
-                {reflections.map((reflection) => {
-                  return (
-                    <ReflectionCard
-                      key={reflection.id}
-                      meeting={newMeeting}
-                      reflection={reflection}
-                    />
-                  )
-                })}
-              </ReflectionGrid>
-            </ReflectionSectionInner>
-          </ScrollableBlock>
-        </ReflectionSection>
-        <TaskSection>
-          <ScrollableBlock>
-            <TaskCardBlock>
-              <MeetingAgendaCards
-                meetingId={meetingId}
-                reflectionGroupId={reflectionGroupId}
-                tasks={tasks}
-                teamId={teamId}
-              />
-            </TaskCardBlock>
-          </ScrollableBlock>
-        </TaskSection>
+        <HeaderContainer>
+          <DiscussHeader>
+            <TopicHeading>
+              <span>{'“'}</span>
+              {`${title}”`}
+            </TopicHeading>
+            <VoteMeta>
+              <VoteIcon name={meetingVoteIcon} />
+              {voteCount}
+            </VoteMeta>
+          </DiscussHeader>
+        </HeaderContainer>
+        <ColumnsContainer>
+          <Column>
+            <LabelContainer>
+              <LabelHeading>
+                {reflections.length} {plural(reflections.length, 'Reflection')}
+              </LabelHeading>
+            </LabelContainer>
+            <Overflow>
+              <ColumnInner>
+                <ReflectionGrid>
+                  {reflections.map((reflection) => {
+                    return (
+                      <ReflectionCard
+                        key={reflection.id}
+                        meeting={newMeeting}
+                        reflection={reflection}
+                      />
+                    )
+                  })}
+                </ReflectionGrid>
+              </ColumnInner>
+            </Overflow>
+          </Column>
+          <TaskColumn>
+            <LabelContainer>
+              <LabelHeading>Takeaway Tasks</LabelHeading>
+            </LabelContainer>
+            <Overflow>
+              <ColumnInner>
+                <TaskCardBlock>
+                  <MeetingAgendaCards
+                    meetingId={meetingId}
+                    reflectionGroupId={reflectionGroupId}
+                    tasks={tasks}
+                    teamId={teamId}
+                  />
+                </TaskCardBlock>
+              </ColumnInner>
+            </Overflow>
+          </TaskColumn>
+        </ColumnsContainer>
       </PhaseWrapper>
       {isFacilitating && (
         <SpacedMeetingControlBar>
@@ -211,6 +242,7 @@ const RetroDiscussPhase = (props: Props) => {
           {!nextStageRes && <ControlButtonBlock />}
         </SpacedMeetingControlBar>
       )}
+      <DiscussHelpMenu floatAboveBottomBar={isFacilitating} />
     </React.Fragment>
   )
 }