+++ src/universal/styles/ui.js
@@ -167,6 +167,8 @@ const TYPE_SEMIBOLD = 600
 
 // Shadows
 // NOTE: levels increase on a scale of 2x
+// #deprecated use elevation.js and named constants instead (TA)
+// https://github.com/ParabolInc/action/issues/2324
 
 const baseShadowColor = [
   makeShadowColor('.15'),
@@ -247,7 +249,6 @@ const ui = {
   // ---------------------------------------------------------------------------
   cardBorderColor: appTheme.palette.mid30l,
   cardBorderRadius: borderRadiusMedium,
-  cardBoxShadow: shadow[0],
   cardButtonHeight: '1.5rem',
   cardButtonBorderColor: appTheme.palette.mid50l,
   cardContentFontSize: '.9375rem',
@@ -484,7 +485,6 @@ const ui = {
   menuBackgroundColor: '#fff',
   menuBorderColor: appTheme.palette.mid30l,
   menuBorderRadius: borderRadiusSmall,
-  menuBoxShadow: shadow[3],
   menuGutterHorizontal: '1rem',
   menuGutterInner: '.75rem',
   menuGutterVertical: '.5rem',
@@ -500,7 +500,6 @@ const ui = {
   // ---------------------------------------------------------------------------
   modalBackdropBackgroundColor: makeShadowColor('.3'),
   modalBorderRadius: borderRadiusLarge,
-  modalBoxShadow: `${shadow[4]}, 0 0 .0625rem rgba(0, 0, 0, .35)`,
   modalButtonSize: BUTTON_SIZE_MEDIUM,
   modalLayoutMain: MODAL_LAYOUT_MAIN,
   modalLayoutMainWithDashAlert: MODAL_LAYOUT_MAIN_WITH_DASH_ALERT,
@@ -536,7 +535,6 @@ const ui = {
 
   // Panels
   // ---------------------------------------------------------------------------
-  panelBoxShadow: shadow[0], // based on dark palette color
   panelInnerBorderColor,
   panelBorderRadius: borderRadiusMedium,
   panelGutter: '1rem',
@@ -617,11 +615,6 @@ const ui = {
   tagHeight: '1rem',
   tagPadding: '0 .5rem',
   tagPalette: ['cool', 'gray', 'midGray', 'light', 'warm', 'yellow', 'blue', 'white'],
-
-  // Tooltips
-  // ---------------------------------------------------------------------------
-  tooltipBorderRadius: borderRadiusSmall,
-  tooltipBoxShadow: shadow[2],
 
   // Transitions
   // ---------------------------------------------------------------------------