+++ webpack/webpack.dev.config.js
@@ -46,7 +46,7 @@ module.exports = {
             cacheDirectory: true,
             babelrc: false,
             plugins: [
-              'syntax-object-rest-spread',
+              'transform-object-rest-spread',
               'syntax-dynamic-import',
               'transform-class-properties',
               ['relay', {artifactDirectory: './src/__generated__'}]